requirements.txt
Django==2.2.11
django-bootstrap-form==3.4
django-crispy-forms==1.8.1
django-environ==0.4.5
django-filter==2.2.0
django-filters==0.2.1
django-import-export==2.0.2
django-plotly-dash==1.1.5
django-redis==4.11.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djongo==1.3.1
ipython==7.12.0
matplotlib==3.1.3
pandas==1.0.1
Pillow==7.0.0
pivottablejs==0.9.0
plotly==4.5.3
prompt-toolkit==3.0.3
Pygments==2.5.2
pymongo==3.10.1









